Handling GET Requests (First GET Requests)
app.get('/', (req, res) => {})
First parameter gets the endpoint or the path (page) to handle GET requests of that page.
Second parameter has the callback function, which is used to send response.
Express take request and response parameters and use it inside any API request.

req is the request, basically whatever comes from the user while making API request.
res is reponse that we are sending to users as a reponse to their request