Handled error of verifyToken method
if(err) next(err), Added this to handle error by verifyToken.

we are calling verifyToken from verifyUser and verifyAdmin. We added above line of code in the third parameter that is the callback function.

verifyUser/admin will be called -> verifyToken is called then -> if error is thrown (For eg: no token) then verifyToken will create error -> without above line of code we were not capturing the error and were directly moving to the callback function in the third parameter when verifyToken was being called....bypassing the error of verifyToken.

This also arose reading undefined values...which gave Internal Server Error 500...Cannot read properties of undefined.