Step 5: Created a component which will subscribe context..
Created a component which will - subscribe context/provide value - to its children, when wrapped around othen components to make them the children.

Children is a special prop, automatically passed to every component, that can be used to render the content included between the opening and closing tags when invoking a component. These kinds of components are identified by the official documentation as “boxes”.

https://codeburst.io/a-complete-guide-to-props-children-in-react-c315fab74e7c

Method 2:
There is another method to subscriibe the context...which is to import Context in the app/parent component and use syntax  <SearchContext.Provider there, and also pass the value arguement by making state in the app/parent component.